update the textures for scene 3
